requirements.txt
appnope==0.1.2
attrs==20.3.0
backcall==0.2.0
branca==0.4.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
cycler==0.10.0
decorator==5.0.7
filelock==3.0.12
Fiona==1.8.19
folium==0.12.1
greenlet==1.1.0
idna==2.10
ipykernel==5.5.3
ipython==7.23.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
matplotlib-inline==0.1.2
msgpack==1.0.2
munch==2.5.0
neovim==0.3.1
numpy==1.20.2
pandas==1.2.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.9.0
pynvim==0.4.3
pyparsing==2.4.7
pyproj==3.0.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
selenium==3.141.0
Shapely==1.7.1
six==1.15.0
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
